+++ packages/automations/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vendoai/automations
 
+## 0.33.0
+
+### Patch Changes
+
+- Updated dependencies [8c7b476]
+- Updated dependencies [9d3f0af]
+  - @vendoai/core@0.33.0
+
 ## 0.32.0
 
 ### Patch Changes